requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
awscli==1.27.155
awscli-local==0.20
boto3==1.26.76
botocore==1.29.76
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.4
docutils==0.16
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
localstack-client==2.2
multidict==6.0.4
pyasn1==0.5.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.1
six==1.16.0
tabulate @ file:///private/tmp/python-tabulate-20221231-6141-6zmrv5/tabulate-0.9.0
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
